requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
diff-match-patch==20121119
Django==1.11.14
django-cors-headers==2.4.0
django-import-export==1.0.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
gunicorn==19.4.5
idna==2.7
invoke==1.1.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.5
paramiko==2.4.1
psutil==5.4.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.3
pycparser==2.18
PyJWT==1.6.4
PyNaCl==1.2.1
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
